refactor: request body의 파라미터를 가져오는 메서드 적용
