+++ tomcat/src/main/java/org/apache/catalina/controller/RegisterController.java
@@ -2,20 +2,18 @@ package org.apache.catalina.controller;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
-import org.apache.catalina.http.HeaderName;
+import org.apache.catalina.HeaderName;
 import org.apache.catalina.http.StatusCode;
 import org.apache.catalina.request.HttpRequest;
-import org.apache.catalina.request.RequestBody;
 import org.apache.catalina.response.HttpResponse;
 
 public class RegisterController extends AbstractController {
 
     @Override
     protected void doPost(HttpRequest request, HttpResponse response) {
-        RequestBody requestBody = request.getBody();
-        String account = requestBody.get("account");
-        String password = requestBody.get("password");
-        String email = requestBody.get("email");
+        String account = request.getBodyParam("account");
+        String password = request.getBodyParam("password");
+        String email = request.getBodyParam("email");
         InMemoryUserRepository.save(new User(account, password, email));
 
         response.setStatusCode(StatusCode.FOUND);
